requirements-dev.txt
black==19.10b0
pylint-django==2.4.2
flake8==3.8.4
flake8-isort==4.0.0
requests-mock==1.8.0
pytest==6.2.2
coverage==5.4
django-coverage-plugin==1.8.0
pytest-django==4.1.0
